Updated documentation structure
