+++ doc/img/neodash-editor.png
