+++ src/card/footer/NeoBarChartFooter.js
@@ -1,7 +1,7 @@
 import React from "react";
-import NeoOptionSelect from "../../component/NeoOptionSelect";
+import NeoOptionSelect from "../../../component/NeoOptionSelect";
 
-class NeoLinePropertySelect extends React.Component {
+class NeoBarPropertySelect extends React.Component {
     constructor(props) {
         super(props);
         this.state = {}
@@ -12,12 +12,12 @@ class NeoLinePropertySelect extends React.Component {
             <div className="neo-property-select" style={{marginLeft: '20px', marginTop: "-8px"}}>
                 <NeoOptionSelect
                     defaultValue={(this.props.propertiesSelected[0]) ? this.props.propertiesSelected[0] : Object.values(this.props.values)[0]}
-                    suffix="-name" label="X-Axis"
+                    suffix="-name" label="Category"
                     onChange={this.props.onChange} value={'name'}
                     options={this.props.categories}/>
                 <NeoOptionSelect
                     defaultValue={(this.props.propertiesSelected[1]) ? this.props.propertiesSelected[1] : Object.values(this.props.values)[1]}
-                    suffix="-name" label="Y-Axis" onChange={this.props.onChange} value={'name'}
+                    suffix="-name" label="Value" onChange={this.props.onChange} value={'name'}
                     options={this.props.values}/>
             </div>
         );
@@ -25,4 +25,4 @@ class NeoLinePropertySelect extends React.Component {
 
 }
 
-export default (NeoLinePropertySelect);
+export default (NeoBarPropertySelect);
