+++ src/card/footer/NeoLineChartFooter.js
@@ -1,7 +1,7 @@
 import React from "react";
-import NeoOptionSelect from "../../component/NeoOptionSelect";
+import NeoOptionSelect from "../../../component/NeoOptionSelect";
 
-class NeoBarPropertySelect extends React.Component {
+class NeoLinePropertySelect extends React.Component {
     constructor(props) {
         super(props);
         this.state = {}
@@ -12,12 +12,12 @@ class NeoBarPropertySelect extends React.Component {
             <div className="neo-property-select" style={{marginLeft: '20px', marginTop: "-8px"}}>
                 <NeoOptionSelect
                     defaultValue={(this.props.propertiesSelected[0]) ? this.props.propertiesSelected[0] : Object.values(this.props.values)[0]}
-                    suffix="-name" label="Category"
+                    suffix="-name" label="X-Axis"
                     onChange={this.props.onChange} value={'name'}
                     options={this.props.categories}/>
                 <NeoOptionSelect
                     defaultValue={(this.props.propertiesSelected[1]) ? this.props.propertiesSelected[1] : Object.values(this.props.values)[1]}
-                    suffix="-name" label="Value" onChange={this.props.onChange} value={'name'}
+                    suffix="-name" label="Y-Axis" onChange={this.props.onChange} value={'name'}
                     options={this.props.values}/>
             </div>
         );
@@ -25,4 +25,4 @@ class NeoBarPropertySelect extends React.Component {
 
 }
 
-export default (NeoBarPropertySelect);
+export default (NeoLinePropertySelect);
