+++ src/card/report/NeoTableReport.js
@@ -1,7 +1,7 @@
 import Table from "react-materialize/lib/Table";
 import React from "react";
 import NeoReport from "../NeoReport";
-import NeoGraphChip from "../../component/NeoGraphChip";
+import NeoGraphChip from "../../../component/NeoGraphChip";
 
 
 class NeoTable extends NeoReport {
